Disable collectDeletedTablesMetrics when config.ExcludeDeleted is true
Hi Grafana team!

We recently set this up via alloy and were surprised to find that the
`collectDeletedTablesMetrics` wasn't disabled when
`config.ExcludeDeleted` was set to true. The query in question takes
north of 20 minutes to execute for us, and we'd like to disable it if
possible.

Happy to make changes if the reviewer would prefer a different
implementation.

Thanks!